[department-of-veterans-affairs/orchid#88] updates local json schema
