+++ src/applications/ask-a-question/tests/data/maximal-test.json
@@ -8,7 +8,7 @@
       "last": "Doe",
       "suffix": "IV"
     },
-    "veteranStatusSection": {
+    "veteranStatus": {
       "veteranStatus": "For the dependent of a Veteran",
       "isDependent": true,
       "relationshipToVeteran": "Daughter",
